# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Bank Statement Analyzer Pro API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /statements:
    get:
      operationId: listStatements
      tags: [statements]
      summary: List uploaded statements
      responses:
        "200":
          description: List of statements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Statement"

    post:
      operationId: uploadStatement
      tags: [statements]
      summary: Upload a bank statement (CSV/XLSX/JSON)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatementUpload"
      responses:
        "201":
          description: Statement uploaded and parsed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Statement"
  /statements/{id}:
    get:
      operationId: getStatement
      tags: [statements]
      summary: Get a statement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Statement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Statement"
        "404":
          description: Not found

    delete:
      operationId: deleteStatement
      tags: [statements]
      summary: Delete a statement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Transactions
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions with optional filters
      parameters:
        - name: statementId
          in: query
          schema:
            type: ["integer", "null"]
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: type
          in: query
          schema:
            type: ["string", "null"]
            enum: [credit, debit, null]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: startDate
          in: query
          schema:
            type: ["string", "null"]
        - name: endDate
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: offset
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Transactions list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionPage"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found

    patch:
      operationId: updateTransaction
      tags: [transactions]
      summary: Update a transaction (category, notes, etc.)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionUpdate"
      responses:
        "200":
          description: Updated transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/bulk-categorize:
    post:
      operationId: bulkCategorize
      tags: [transactions]
      summary: Auto-categorize all uncategorized transactions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkCategorizeInput"
      responses:
        "200":
          description: Categorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkCategorizeResult"

  # Categories
...
  /users/profile:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    patch:
      operationId: updateUserProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # Admin
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # Auth
...
    User:
      type: object
      required: [id, email, name, role, plan]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [user, admin, super_admin]
        plan:
          type: string
          enum: [free, pro, enterprise]
        status:
          type: string
          enum: [active, suspended]
        createdAt:
          type: string
    UserProfileUpdate:
      type: object
      properties:
        name:
          type: string
        currency:
          type: ["string", "null"]

    # Statements
    Statement:
      type: object
      required: [id, bankName, fileName, transactionCount, totalDebits, totalCredits, startDate, endDate, createdAt]
      properties:
        id:
          type: integer
        bankName:
          type: string
        fileName:
          type: string
        transactionCount:
          type: integer
        totalDebits:
          type: number
        totalCredits:
          type: number
        startDate:
          type: string
        endDate:
          type: string
        status:
          type: string
          enum: [processing, ready, error]
        createdAt:
          type: string
    StatementUpload:
      type: object
      required: [bankName, transactions]
      properties:
        bankName:
          type: string
        fileName:
          type: string
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/TransactionImport"
    TransactionImport:
      type: object
      required: [date, description, amount, type]
      properties:
        date:
          type: string
        description:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [credit, debit]
        merchant:
          type: ["string", "null"]
        reference:
          type: ["string", "null"]

    # Transactions
    Transaction:
      type: object
      required: [id, statementId, date, description, amount, type]
      properties:
        id:
          type: integer
        statementId:
          type: integer
        date:
          type: string
        description:
          type: string
        merchant:
          type: ["string", "null"]
        amount:
          type: number
        type:
          type: string
          enum: [credit, debit]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isRecurring:
          type: boolean
        createdAt:
          type: string
    TransactionPage:
      type: object
      required: [transactions, total]
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
    TransactionUpdate:
      type: object
      properties:
        categoryId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        merchant:
          type: ["string", "null"]
        isRecurring:
          type: ["boolean", "null"]
    BulkCategorizeInput:
      type: object
      properties:
        statementId:
          type: ["integer", "null"]
    BulkCategorizeResult:
      type: object
      required: [updated]
      properties:
        updated:
          type: integer

    # Categories
...